Update upload product image, allow .webp format, and fix deleteFile function to check if file exists before deleting

+++ src/utils/deleteFile.ts
@@ -1,7 +1,7 @@
 import fs from "fs"
 
 export const deleteFile = (filePath:string)=>{
-  fs.unlink(filePath,(err)=>{
+  fs.existsSync(filePath) && fs.unlink(filePath,(err)=>{
     if(err) throw err
   })
 }
